splitted the game class and the player classes in seperate files
