+++ app/src/types/GameEngine.d.ts
@@ -25,3 +25,9 @@ interface GameState {
 	objects: GameObject[];
 }
 
+interface AIBrainData {
+	aiLastBallDistance: number;
+	aiDelayCounter: number;
+	aiLastTargetParam: number;
+	lastAIMovementDirection: number;
+}
